update CUDA_TRY macross to explicitly cast cuda status to int
This change is needed for compatibility with compile-time checks in fmt v10.1.1

+++ operators/orsi/orsi_segmentation_preprocessor/segmentation_preprocessor.cpp
@@ -46,7 +46,7 @@
                     __LINE__,                                                              \
                     __FILE__,                                                              \
                     cudaGetErrorString(_holoscan_cuda_err),                                \
-                    _holoscan_cuda_err);                                                   \
+                    static_cast<int>(_holoscan_cuda_err));                                 \
     }                                                                                      \
     _holoscan_cuda_err;                                                                    \
   })
